upgrade tests
